requirements.txt
aiohttp==3.10.11
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
bidict==0.22.1
blinker==1.7.0
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.17
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
ddtrace
fastapi>=0.109.2
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
grpcio==1.65.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.1
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain-chroma==0.1.4
langchain-core
langchain-text-splitters
langchain==0.3.7
langchain-openai==0.2.3
langchain-community==0.3.3
langsmith<0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.52.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
overrides==7.7.0
pandas==2.2.2
posthog==3.5.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
PyJWT==2.8.0
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-engineio==4.8.2
python-multipart==0.0.18
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.31
starlette==0.40.0
setuptools==70.0.0
sympy==1.13.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.3
twilio==8.12.0
typer==0.12.3
typing-inspect==0.9.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
wrapt==1.16.0
wsproto==1.2.0
yarl>=1.12.0
zipp==3.19.1
pymupdf==1.24.7
soundfile==0.12.1
librosa==0.10.2
python-json-logger==2.0.7
sse-starlette==2.1.2
neo4j==5.23.1
beautifulsoup4==4.12.3
inflection==0.5.1
ddtrace==2.13.0
livekit==0.18.0
livekit_agents==0.11.1
livekit_plugins_openai==0.10.5
livekit_plugins_silero==0.7.3
